iPhone: Fetching watering info is now complete
The data is modeled in waterings.m, and a very brief summary of each waterig is displayed in the main view.

Next step: Allow the user to view details of and edit waterings.

Next next step: Give each watering a user-editable name (cause the present method of showing a derived summary sucks). This will hopefully happen at the same time that zones each get a human editable name.