iPhone: Add period editing
For "every N days" waterings, you can now set the period (in days).

Also fixed the broken tap/drag behavior of the date/time pickers in the action sheets by using this examle:

http://stackoverflow.com/questions/1262574/add-uipickerview-a-button-in-action-sheet-how

I finally realized that I wanted to embed a UIPickerView inside a UIActionSheet (which is the more generic solution that UIDatePicker), and so I finally found some decent advice on how to do this. There is a lot of bad iOS development advice out there.